refactored input and others

+++ Algorithm.hpp
@@ -6,7 +6,7 @@ class Algorithm
 {
 public:
 	typedef std::vector<std::vector< std::list<int>* > >* RouteContainer;
-	static RouteContainer computeFloydWarshall(int** distance, int n);
+	static RouteContainer computeFloydWarshall(double** distance, int n);
 	~Algorithm(void);
 private:
 	static std::list<int>* Path(const std::vector<std::vector<int> >&,int,int);
